Recorded activity displays after entering/exiting activity screen
There was a bug where a recorded activity would initially appear,
then if you backed out of the activity screen, then re-entered it,
it would permanentely dissapear. This occurred because certain keys
that were to not be displayed (ID, name, start, end), were simply
deleted. The start key was needed for another function outside of
displaying the recorded activity data, so deleting it caused the bug.